+++ src/noisy_channel.cpp
@@ -17,8 +17,8 @@ void binary_symmetric(std::array<uint64_t,COLS> &codeword){
   std::bernoulli_distribution distribution(0.1); // Noise Level p=0.1
 
   bitset<64> noise;
-  for (int i = 0; i < COLS; i++){ 
-    for (int j = 0; j < 64; j++){ //64-Bit Noise Sequenzen werden erzeugt 
+  for (size_t i = 0; i < COLS; i++){ 
+    for (size_t j = 0; j < 64; j++){ //64-Bit Noise Sequenzen werden erzeugt 
         noise[j] = distribution(generator);
     } 
     //cout << "Noise: " << bitset<64>(noise) << endl;
@@ -58,10 +58,10 @@ void GaussianNoise(std::array<uint64_t, COLS> &codeword,
     //std::cout << t.size() << "Signal VEctor Size (gemappt) " << std::endl;
 
     int count = 0;
-    for (int i = 0; i<COLS; i++){
+    for (size_t i = 0; i<COLS; i++){
         //std::cout << bitset<64>(codeword[i]) << std::endl;
       bitset<64> current_codeword = bitset<64>(codeword[i]);
-      for (int j = 0; j < SCALE; j++){ // 512 - codeword groesse
+      for (size_t j = 0; j < SCALE; j++){ // 512 - codeword groesse
         current_codeword[j] == 0? t[SCALE*i+j]=-a : t[SCALE*i+j]=a; // KORREKTUR codeword indizes separat vorberechnen
         //std::cout << count <<  " - codeword[i]" << (current_codeword)[j] << " " << "t[i]:" << t[SCALE*i+j] << std::endl;
         count++;
@@ -77,7 +77,7 @@ void GaussianNoise(std::array<uint64_t, COLS> &codeword,
     //std::array<float, COLS*SCALE> r = {}; // recieved vector
     //std::cout << "Recieved Vector Size " << r.size() << std::endl;
     
-    for (int i = 0; i<t.size(); i++){
+    for (size_t i = 0; i<t.size(); i++){
         
         r[i] = t[i] + distribution(generator);
         //std::cout << t[i] << " " << r[i] << std::endl;
