+++ src/preprocessing.cpp
@@ -11,23 +11,23 @@ void fillMatrix (int8_t (&base)[ROWS][COLS], std::mt19937 &generator){
 
   // HAT WAS MIT CODE RATE ZU TUN ABER ERSTMAL HARDCODED!!!!!!!!!!!
     // für DUAL DIAGONAL FORM FÜLLEN
-    for (uint8_t i = 0; i < ROWS; i++){
-        for (uint8_t j = 4; j < COLS; j++) {
+    for (size_t i = 0; i < ROWS; i++){
+        for (size_t j = 4; j < COLS; j++) {
             ((i == (j-4)) or (i == (j - 4 + 1))) ? base[i][j] = 0 : base[i][j] = -1;
         }
     }
     
     //--- "A" TEIL DER BASE MATRIX MIT ZUFÄLLIGEN WERTEN FÜLLEN
-    for (uint8_t i = 0; i < 4; i++){
-        for (uint8_t j = 0; j < 4; j++) {
+    for (size_t i = 0; i < 4; i++){
+        for (size_t j = 0; j < 4; j++) {
             base[i][j] = distribution(generator);  
         }
     }
 
     cout << " Ausgabe Matrix " << endl;
-    for (uint8_t i = 0; i < 4; i++){
+    for (size_t i = 0; i < 4; i++){
         cout << " " << endl;
-        for (uint8_t j = 0; j < 8; j++) {
+        for (size_t j = 0; j < 8; j++) {
             std::cout << +base[i][j] << " ";
             
         }
@@ -38,11 +38,11 @@ bool girthCheck ( int8_t (&base) [ROWS][COLS]){
    //std::cout << " Girth-4 Check beginnt... "<< std::endl;
     //cout << +COLS << +ROWS << endl;
 
-    for (uint8_t step_c = 1; step_c < COLS; step_c++) {
+    for (size_t step_c = 1; step_c < COLS; step_c++) {
       
-      for (uint8_t step_r = 1; step_r < ROWS; step_r++) {
-        for ( uint8_t i = 0; i < ROWS - step_r; i++ ){
-            for (uint8_t j = 0; j < COLS - step_c; j++){
+      for (size_t step_r = 1; step_r < ROWS; step_r++) {
+        for ( size_t i = 0; i < ROWS - step_r; i++ ){
+            for (size_t j = 0; j < COLS - step_c; j++){
               if (base[i][j] != -1 and base[i+step_r][j] != -1 and base[i][j+step_c] != -1 and base[i+step_r][j+step_c] != -1){
                 // das mit Modulo 64 dokumentieren, z:B wieso denn "+64"?
